requirements.txt
fastapi[all]
uvicorn[standard]
onnx==1.10.2
pillow==9.1.1
onnxruntime==1.11.1
protobuf==3.20.0
scikit-learn==1.0.2
pytest==7.1.2
matplotlib==3.4.3
opencv-python-headless